# admin/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_admin-template_backend_internal_dto.AssignMenusReq:
    properties:
      menu_ids:
        items:
          type: integer
        type: array
    required:
    - menu_ids
    type: object
  github_com_admin-template_backend_internal_dto.BatchDeleteReq:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  github_com_admin-template_backend_internal_dto.CreateRoleReq:
    properties:
      code:
        maxLength: 64
        minLength: 1
        type: string
      description:
        maxLength: 255
        type: string
      menu_ids:
        items:
          type: integer
        type: array
      name:
        maxLength: 64
        minLength: 1
        type: string
      sort:
        type: integer
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - code
    - name
    type: object
...
  github_com_admin-template_backend_internal_dto.RoleResp:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menu_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      sort:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_admin-template_backend_internal_dto.UpdateRoleReq:
    properties:
      description:
        maxLength: 255
        type: string
      menu_ids:
        items:
          type: integer
        type: array
      name:
        maxLength: 64
        type: string
      sort:
        type: integer
      status:
        enum:
        - 1
        - 2
        type: integer
    type: object
...
  github_com_admin-template_backend_internal_pkg_response.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_admin-template_backend_internal_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: RBAC admin backend — auth, users/roles/menus CRUD, operation log, file
    upload, Excel/CSV export.
  title: Admin Template API
  version: "1.0"
paths:
...
  /system/roles:
    get:
      parameters:
      - description: 页码 (默认 1)
        in: query
        name: page
        type: integer
      - description: 每页大小 (默认 10, 最大 200)
        in: query
        name: size
        type: integer
      - description: 模糊搜索名称/编码
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - System/Roles
    post:
      consumes:
      - application/json
      parameters:
      - description: 新角色
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.CreateRoleReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.RoleResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
        "409":
          description: 角色编码已存在
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - System/Roles
  /system/roles/{id}:
    delete:
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 删除单个角色
      tags:
      - System/Roles
    get:
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.RoleResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 获取单个角色
      tags:
      - System/Roles
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要更新的字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.UpdateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.RoleResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - System/Roles
  /system/roles/{id}/menus:
    post:
      consumes:
      - application/json
      description: 覆盖式写入；传入空数组可清空。
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.AssignMenusReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 给角色分配菜单
      tags:
      - System/Roles
  /system/roles/batch-delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.BatchDeleteReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 批量删除角色
      tags:
      - System/Roles
  /system/roles/export/csv:
    get:
      parameters:
      - description: 模糊搜索
        in: query
        name: keyword
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出角色列表为 CSV
      tags:
      - System/Roles
  /system/roles/export/excel:
    get:
      parameters:
      - description: 模糊搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出角色列表为 Excel
      tags:
      - System/Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"